requirements.txt
PyQt5-Qt5>=5.15.2
PyQt5-sip>=12.11.0
pyqt5>=5.15.7
pyyaml>=6.0
greenlet>=1.1.3.post0
sqlalchemy>=1.4.41
PyMySQL>=1.0.2
